mssc-beer-service/src/test/java/balaji/springframweork/msscbeerservice/web/controller/BeerControllerTest.java
package balaji.springframweork.msscbeerservice.web.controller;

import balaji.springframweork.msscbeerservice.web.model.BeerDTO;
import balaji.springframweork.msscbeerservice.web.model.BeerStyleEnum;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import java.math.BigDecimal;
import java.util.UUID;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@WebMvcTest(BeerController.class)
class BeerControllerTest {

    @Autowired
    private MockMvc mockMvc;
    @Autowired
    private ObjectMapper objectMapper;

    @Test
    void getBeerById() throws Exception {
        mockMvc.perform(get("/api/v1/beer/" + UUID.randomUUID())
                .accept(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk());
    }

    @Test
    void saveNewBeer() throws Exception {
        BeerDTO beerDTO = BeerDTO.builder()
                .beerName("Sample")
                .beerStyle(BeerStyleEnum.ALE)
                .upc(4487219797L)
                .price(new BigDecimal("11.95"))
                .build();
        String beerDTOJSON = objectMapper.writeValueAsString(beerDTO);

        mockMvc.perform(post("/api/v1/beer")
                .contentType(MediaType.APPLICATION_JSON)
                .content(beerDTOJSON))
                .andExpect(status().isCreated());
    }

    @Test
    void updateBeer() throws Exception {
        BeerDTO beerDTO = BeerDTO.builder()
                .beerName("Sample")
                .beerStyle(BeerStyleEnum.ALE)
                .upc(4487219797L)
                .price(new BigDecimal("11.95"))
                .build();
        String beerDTOJSON = objectMapper.writeValueAsString(beerDTO);

        mockMvc.perform(put("/api/v1/beer/" + UUID.randomUUID())
                .contentType(MediaType.APPLICATION_JSON)
                .content(beerDTOJSON))
                .andExpect(status().isNoContent());
    }
}